requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
langgraph==0.1.5
langchain==0.2.11
langchain-groq==0.1.6
python-dotenv==1.0.1
pydantic==2.8.2
